DataCommand Class
the DataCommand class extends Command and with added functionality of
easily writing values to a buffered reader to be saved as a csv on a
flashdrive inserted into the RoboRio. As far as implementation is
concerned anything that extends the DataCommand class gets the name
"-DataCommand" (i.e. VelocityDriveDataCommand). The DataCommand class
has documentation if you want to know what's going on